pyproject.toml
[tool.poetry]
name = "nonebot_plugin_randomtkk"
version = "0.1.0"
description = "Find Tan Kuku!"
authors = ["KafCoppelia <k740677208@gmail.com>"]
license = "MIT"
packages = [{include = "nonebot_plugin_randomtkk"}]

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0-beta.2"
nonebot-adapter-onebot = "^2.0.0-beta.1"
pillow = "^9.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "tencentyun"
url = "https://mirrors.tencent.com/pypi/simple"
default = true